refactor: Version 2 of my project submission to Udacity
There was a problem with the say I was replacing missing values with
NaNs. I rewrote the code to correctly change all the missing values by
NaNs which allowed me to do the data cleaning properly.

This changed the dataset a bit and I ended up with more columns. This
meant that when I did PCA, I had different coefficients for my columns.
I also had to update which columns had the most weight in my model.

Finally I updated some visuals as well.